notification-service/requirements.txt
fastapi
uvicorn
requests
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
